oci/tests: Ensure infra env is stopped at the end
Update tftestenv with signal handling for graceful terraform shutdown.

Move the environment stop into a defer block to ensure that the
environment is always stopped at the end if -retain is not configured.
This will make sure that the environment is deleted in case any program
crashed, especially in CI environments.

Signed-off-by: Sunny <darkowlzz@protonmail.com>